Fix handing id when loading from path (#332)
* Fix handing id when loading from path

Previously, the id for the pyg data object would become the entire (possibly length) path to the pdb file OR only the filename + chain ID. However, in the later case, the '.pdb' extension would still show up under the id field.

This commit fixes the above to be consistent and to not contain '.pdb' extension in the pyg data id.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add non-standard Cystine

From Wikipedia: "The thiol is susceptible to oxidation to give the disulfide derivative cystine, which serves an important structural role in many proteins. In this case, the symbol Cyx is sometimes used. The deprotonated form can generally be described by the symbol Cym as well."

* Update CHANGELOG.md

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>